% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvi.R
\name{bvi}
\alias{bvi}
\title{Biological Value Index}
\usage{
bvi(data, p = 0.95, other = T)
}
\arguments{
\item{data}{an object of class matrix or data.frame. rownames must be samples and colnames must be species (or any other taxa).}

\item{p}{The cutoff percentage to use, expressed relative to 1 (i.e. for a cutoff of 95% use p = 0.95)}
}
\value{
results An object of class data.frame containing a column with species name (spp), the scores for each species in each sample, a column (BVI) with the Biological Value Index, and a column (rBVI) with the Relative Biological Value Index.
}
\description{
Calculates the Biological Value Index (Sanders, 1960) with modifications from Loya-Salina & Escofet (1990). It also provides a relative index, that allows for comparisons.
}
\author{
Villasenor-Derbez, J.C.
}
\seealso{
bvi_plot
}

