% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvi_plot.R
\name{bvi_plot}
\alias{bvi_plot}
\title{Biological Value Index grapher}
\usage{
bvi_plot(bvi_scores)
}
\arguments{
\item{bvi_scores}{An object of class data.frame generated generated by bvi()}
}
\description{
Creates a graph to graphicaly represent the score matrix of the Biological Value Index.
}
\seealso{
bvi
}

